More stuff.

+++ com/mraof/minestuck/util/SburbConnection.java
@@ -7,7 +7,9 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.TreeMap;
 
 import com.mraof.minestuck.Minestuck;
@@ -19,9 +21,9 @@ import net.minecraft.world.World;
 public class SburbConnection {
 
 	private static TreeMap<String,ComputerData> serversOpen = new TreeMap<String,ComputerData>();
-	private static ArrayList<IConnectionListener> listeners = new ArrayList<IConnectionListener>();
+	private static List<IConnectionListener> listeners = Collections.synchronizedList(new ArrayList<IConnectionListener>());
 	private static ArrayList<SburbConnection> connections = new ArrayList<SburbConnection>();
-	private static HashMap<String,ComputerData> mainConnections = new HashMap<String,ComputerData>();
+	private static HashMap<String,ComputerData> mainConnections = new HashMap<String,ComputerData>();	//Connections waiting for
 
 	public static ArrayList<String> getServersOpen() {
 		return new ArrayList<String>(serversOpen.keySet());
@@ -34,14 +36,57 @@ public class SburbConnection {
 		return null;
 	}
 	
+	public static boolean isResuming(String player, boolean isClient){
+		for(ComputerData data : mainConnections.values())
+			if(data.owner.equals(player) && data.isClient == isClient)
+				return true;
+		return false;
+				
+	}
+	
+	public static void resumeConnection(String player, int x, int y, int z, int dimensionId, boolean isClient){
+		for(SburbConnection c : connections)
+			if(c.isMain && (isClient && c.clientName.equals(player) || !isClient && c.serverName.equals(player)))
+				if(isClient){
+					ComputerData data = mainConnections.get(c.serverName);
+					if(data == null)
+						mainConnections.put(player, new ComputerData(player, x, y, z, dimensionId,isClient));
+					else{
+						c.server = data;
+						mainConnections.remove(c.serverName);
+						c.client = new ComputerData(player, x, y, z, dimensionId,isClient);
+						synchronized(listeners){
+							for(IConnectionListener listener : listeners)
+								listener.onConnected(player,c.serverName);
+						}
+					}
+				}else{
+					ComputerData data = mainConnections.get(c.clientName);
+					if(data == null)
+						mainConnections.put(player, new ComputerData(player, x, y, z, dimensionId,isClient));
+					else{
+						c.client = data;
+						mainConnections.remove(c.clientName);
+						c.server = new ComputerData(player, x, y, z, dimensionId,isClient);
+						synchronized(listeners){
+							for(IConnectionListener listener : listeners)
+								listener.onConnected(c.clientName,player);
+						}
+					}
+				}
+	}
+	
 	public static boolean hasMainClient(String client){ //Returns true if a main connection has a client of the specified name
 		for(SburbConnection c : connections)
-			if(c.clientName.equals(client) || c.client != null && c.client.owner.equals(client))
+			if(c.isMain && (c.clientName.equals(client) || c.client != null && c.client.owner.equals(client)))
 				return true;
 		return false;
 	}
 	
 	public static boolean hasMainServer(String server){
+		for(SburbConnection c : connections)
+			if(c.isMain && (c.serverName.equals(server) || c.server != null && c.server.owner.equals(server)))
+				return true;
 		return false;
 	}
 	
@@ -50,10 +95,12 @@ public class SburbConnection {
 		if(serversOpen.containsKey(player))
 			return;
 		
-		serversOpen.put(player,new ComputerData(player,x,y,z,dimension));
+		serversOpen.put(player,new ComputerData(player,x,y,z,dimension,false));
 		
-		for (Object listener : listeners) {
-			((IConnectionListener)listener).onServerOpen(player);
+		synchronized(listeners){
+			for (IConnectionListener listener : listeners) {
+				listener.onServerOpen(player);
+			}
 		}
 	}
 	
@@ -66,19 +113,18 @@ public class SburbConnection {
 				return;
 			}
 			
-			connections.add(new SburbConnection(new ComputerData(client,x,y,z,dimension),serversOpen.remove(server)));
+			connections.add(new SburbConnection(new ComputerData(client,x,y,z,dimension,true),serversOpen.remove(server)));
 			if(MinestuckSaveHandler.lands.contains((byte) dimension))
 				connections.get(connections.size()-1).enteredGame = true;
-			for (Object listener : listeners) {
-				((IConnectionListener)listener).onConnected(client,server);
+			synchronized(listeners){
+				for (IConnectionListener listener : listeners)
+					listener.onConnected(client,server);
 			}
 		}
 		else Debug.print("Connection denied, the specific server wasn't open, server:"+server);
 	}
 	
 	public static void connectionClosed(String client, String server){
-		for(IConnectionListener listener : listeners)
-			listener.onConnectionClosed(client, server);
 		if(client.isEmpty())
 			serversOpen.remove(server);
 		else
@@ -92,14 +138,20 @@ public class SburbConnection {
 					}else connections.remove(connect);
 					break;
 				}
+		synchronized(listeners){
+			for(IConnectionListener listener : listeners)
+				listener.onConnectionClosed(client, server);
+		}
 	}
 	
 	public static boolean giveItems(String client){
 		for(SburbConnection connect : connections)
 			if(connect.client != null && connect.client.owner.equals(client) && !connect.isMain && !connect.enteredGame){
 				connect.isMain = true;
-				for(IConnectionListener listener : listeners)
-					listener.newPermaConnection(client, connect.server.owner);
+				synchronized(listeners){
+					for(IConnectionListener listener : listeners)
+						listener.newPermaConnection(client, connect.server.owner);
+				}
 				return true;
 			}
 		return false;
@@ -114,36 +166,43 @@ public class SburbConnection {
 	}
 	
 	public static void addListener(IConnectionListener listener) {
-		if(!listeners.contains(listener))
-			listeners.add(listener);
+		synchronized(listeners){
+			if(!listeners.contains(listener))
+				listeners.add(listener);
+		}
 	}
 	
 	public static void removeListener(IConnectionListener listener){
-		listeners.remove(listener);
+		synchronized(listeners){
+			listeners.remove(listener);
+		}
 	}
 	
 	public static void saveData(File file){
-		try{
-			DataOutputStream stream = new DataOutputStream(new FileOutputStream(file));
-			for(SburbConnection c : connections){
-				if(c.client != null && c.server != null){
-					stream.writeBoolean(true);
-					c.client.save(stream);
-					c.server.save(stream);
-					stream.writeBoolean(c.isMain);
-					if(c.isMain)
+		if(file.exists()){
+			try{
+				DataOutputStream stream = new DataOutputStream(new FileOutputStream(file));
+				for(SburbConnection c : connections){
+					if(c.client != null && c.server != null){
+						stream.writeBoolean(true);
+						c.client.save(stream);
+						c.server.save(stream);
+						stream.writeBoolean(c.isMain);
+						if(c.isMain)
+							stream.writeBoolean(c.enteredGame);
+					}
+					else{
+						stream.writeBoolean(false);
+						stream.write((c.clientName+"\n").getBytes());
+						stream.write((c.serverName+"\n").getBytes());
 						stream.writeBoolean(c.enteredGame);
+					}
 				}
-				else{
-					stream.write((c.clientName+"\n").getBytes());
-					stream.write((c.serverName+"\n").getBytes());
-					stream.writeBoolean(c.enteredGame);
-				}
+				stream.close();
+				Debug.print(connections.size()+" connection(s) saved,"+stream.size());
+			} catch(IOException e){
+				e.printStackTrace();
 			}
-			stream.close();
-			Debug.print(connections.size()+" connections saved");
-		} catch(IOException e){
-			e.printStackTrace();
 		}
 	}
 	
@@ -151,7 +210,7 @@ public class SburbConnection {
 		if(file.exists()){
 			try{
 				DataInputStream stream = new DataInputStream(new FileInputStream(file));
-				connections.clear();
+				connections.clear();Debug.print(stream.available());
 				while(stream.available() > 0){
 					boolean connected = stream.readBoolean();
 					SburbConnection c = new SburbConnection(null, null);
@@ -172,7 +231,7 @@ public class SburbConnection {
 					connections.add(c);
 				}
 				stream.close();
-				Debug.print(connections.size()+" connections loaded");
+				Debug.print(connections.size()+" connection(s) loaded");
 				checkConnections();
 			}catch(IOException e){
 				e.printStackTrace();
@@ -203,7 +262,8 @@ public class SburbConnection {
 		private int x, y, z;
 		private int dimension;
 		private String owner;
-		private ComputerData(String owner,int x,int y,int z,int dimension){
+		private boolean isClient;
+		private ComputerData(String owner,int x,int y,int z,int dimension,boolean isClient){
 			this.owner = owner;
 			this.x = x;
 			this.y = y;
@@ -237,9 +297,9 @@ public class SburbConnection {
 	//Non static stuff
 	
 	private ComputerData client;
-	private String clientName;
+	private String clientName = "";
 	private ComputerData server;
-	private String serverName;
+	private String serverName = "";
 	private boolean isMain;
 	private boolean enteredGame;
 	
@@ -250,6 +310,16 @@ public class SburbConnection {
 	
 	public ComputerData getClient(){return client;}
 	public ComputerData getServer(){return server;}
+	public String getClientName(){
+		if(clientName.isEmpty())
+			return client.owner;
+		else return clientName;
+	}
+	public String getServerName(){
+		if(serverName.isEmpty())
+			return server.owner;
+		else return serverName;
+	}
 	public boolean givenItems(){return isMain;}
 	public boolean enteredGame(){return enteredGame;}
